Improve the limit adjustment logic and add safety checks
If limits are calculated for f64 data that uses a RAT_FUNC compu method,
then the ration function will be applied to the input values f64::MIN
and f64::MAX. In this situation it is essential to divide first and then multiply.
Even then the result can be +/-INF, so a safety check has been added.